HDC - Se crea calculo y se guarda en base de datos

+++ Modules/HDC/Database/Migrations/2023_09_03_200135_create_family_person_footprints_table.php
@@ -16,10 +16,10 @@ class CreateFamilyPersonFootprintsTable extends Migration
         Schema::create('family_person_footprints', function (Blueprint $table) {
             $table->id();
             $table->foreignId('person_id')->constrained()->onDelete('cascade');
-            $table->integer('carbon_print');
+            $table->decimal('carbon_print');
             $table->softDeletes();
             $table->timestamps();
-           
+
 
         });
     }
